update to version 12.0.0 (#74)

+++ src/main/java/one/talon/model/AccessLogEntry.java
@@ -37,7 +37,7 @@ public class AccessLogEntry {
 
   public static final String SERIALIZED_NAME_STATUS = "status";
   @SerializedName(SERIALIZED_NAME_STATUS)
-  private Integer status;
+  private Long status;
 
   public static final String SERIALIZED_NAME_METHOD = "method";
   @SerializedName(SERIALIZED_NAME_METHOD)
@@ -82,7 +82,7 @@ public class AccessLogEntry {
   }
 
 
-  public AccessLogEntry status(Integer status) {
+  public AccessLogEntry status(Long status) {
     
     this.status = status;
     return this;
@@ -94,12 +94,12 @@ public class AccessLogEntry {
   **/
   @ApiModelProperty(example = "200", required = true, value = "HTTP status code of response.")
 
-  public Integer getStatus() {
+  public Long getStatus() {
     return status;
   }
 
 
-  public void setStatus(Integer status) {
+  public void setStatus(Long status) {
     this.status = status;
   }
