+++ src/main/java/one/talon/model/AccountAnalytics.java
@@ -32,86 +32,86 @@ import org.threeten.bp.OffsetDateTime;
 public class AccountAnalytics {
   public static final String SERIALIZED_NAME_APPLICATIONS = "applications";
   @SerializedName(SERIALIZED_NAME_APPLICATIONS)
-  private Integer applications;
+  private Long applications;
 
   public static final String SERIALIZED_NAME_LIVE_APPLICATIONS = "liveApplications";
   @SerializedName(SERIALIZED_NAME_LIVE_APPLICATIONS)
-  private Integer liveApplications;
+  private Long liveApplications;
 
   public static final String SERIALIZED_NAME_SANDBOX_APPLICATIONS = "sandboxApplications";
   @SerializedName(SERIALIZED_NAME_SANDBOX_APPLICATIONS)
-  private Integer sandboxApplications;
+  private Long sandboxApplications;
 
   public static final String SERIALIZED_NAME_CAMPAIGNS = "campaigns";
   @SerializedName(SERIALIZED_NAME_CAMPAIGNS)
-  private Integer campaigns;
+  private Long campaigns;
 
   public static final String SERIALIZED_NAME_ACTIVE_CAMPAIGNS = "activeCampaigns";
   @SerializedName(SERIALIZED_NAME_ACTIVE_CAMPAIGNS)
-  private Integer activeCampaigns;
+  private Long activeCampaigns;
 
   public static final String SERIALIZED_NAME_LIVE_ACTIVE_CAMPAIGNS = "liveActiveCampaigns";
   @SerializedName(SERIALIZED_NAME_LIVE_ACTIVE_CAMPAIGNS)
-  private Integer liveActiveCampaigns;
+  private Long liveActiveCampaigns;
 
   public static final String SERIALIZED_NAME_COUPONS = "coupons";
   @SerializedName(SERIALIZED_NAME_COUPONS)
-  private Integer coupons;
+  private Long coupons;
 
   public static final String SERIALIZED_NAME_ACTIVE_COUPONS = "activeCoupons";
   @SerializedName(SERIALIZED_NAME_ACTIVE_COUPONS)
-  private Integer activeCoupons;
+  private Long activeCoupons;
 
   public static final String SERIALIZED_NAME_EXPIRED_COUPONS = "expiredCoupons";
   @SerializedName(SERIALIZED_NAME_EXPIRED_COUPONS)
-  private Integer expiredCoupons;
+  private Long expiredCoupons;
 
   public static final String SERIALIZED_NAME_REFERRAL_CODES = "referralCodes";
   @SerializedName(SERIALIZED_NAME_REFERRAL_CODES)
-  private Integer referralCodes;
+  private Long referralCodes;
 
   public static final String SERIALIZED_NAME_ACTIVE_REFERRAL_CODES = "activeReferralCodes";
   @SerializedName(SERIALIZED_NAME_ACTIVE_REFERRAL_CODES)
-  private Integer activeReferralCodes;
+  private Long activeReferralCodes;
 
   public static final String SERIALIZED_NAME_EXPIRED_REFERRAL_CODES = "expiredReferralCodes";
   @SerializedName(SERIALIZED_NAME_EXPIRED_REFERRAL_CODES)
-  private Integer expiredReferralCodes;
+  private Long expiredReferralCodes;
 
   public static final String SERIALIZED_NAME_ACTIVE_RULES = "activeRules";
   @SerializedName(SERIALIZED_NAME_ACTIVE_RULES)
-  private Integer activeRules;
+  private Long activeRules;
 
   public static final String SERIALIZED_NAME_USERS = "users";
   @SerializedName(SERIALIZED_NAME_USERS)
-  private Integer users;
+  private Long users;
 
   public static final String SERIALIZED_NAME_ROLES = "roles";
   @SerializedName(SERIALIZED_NAME_ROLES)
-  private Integer roles;
+  private Long roles;
 
   public static final String SERIALIZED_NAME_CUSTOM_ATTRIBUTES = "customAttributes";
   @SerializedName(SERIALIZED_NAME_CUSTOM_ATTRIBUTES)
-  private Integer customAttributes;
+  private Long customAttributes;
 
   public static final String SERIALIZED_NAME_WEBHOOKS = "webhooks";
   @SerializedName(SERIALIZED_NAME_WEBHOOKS)
-  private Integer webhooks;
+  private Long webhooks;
 
   public static final String SERIALIZED_NAME_LOYALTY_PROGRAMS = "loyaltyPrograms";
   @SerializedName(SERIALIZED_NAME_LOYALTY_PROGRAMS)
-  private Integer loyaltyPrograms;
+  private Long loyaltyPrograms;
 
   public static final String SERIALIZED_NAME_LIVE_LOYALTY_PROGRAMS = "liveLoyaltyPrograms";
   @SerializedName(SERIALIZED_NAME_LIVE_LOYALTY_PROGRAMS)
-  private Integer liveLoyaltyPrograms;
+  private Long liveLoyaltyPrograms;
 
   public static final String SERIALIZED_NAME_LAST_UPDATED_AT = "lastUpdatedAt";
   @SerializedName(SERIALIZED_NAME_LAST_UPDATED_AT)
   private OffsetDateTime lastUpdatedAt;
 
 
-  public AccountAnalytics applications(Integer applications) {
+  public AccountAnalytics applications(Long applications) {
     
     this.applications = applications;
     return this;
@@ -123,17 +123,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "11", required = true, value = "Total number of applications in the account.")
 
-  public Integer getApplications() {
+  public Long getApplications() {
     return applications;
   }
 
 
-  public void setApplications(Integer applications) {
+  public void setApplications(Long applications) {
     this.applications = applications;
   }
 
 
-  public AccountAnalytics liveApplications(Integer liveApplications) {
+  public AccountAnalytics liveApplications(Long liveApplications) {
     
     this.liveApplications = liveApplications;
     return this;
@@ -145,17 +145,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "6", required = true, value = "Total number of live applications in the account.")
 
-  public Integer getLiveApplications() {
+  public Long getLiveApplications() {
     return liveApplications;
   }
 
 
-  public void setLiveApplications(Integer liveApplications) {
+  public void setLiveApplications(Long liveApplications) {
     this.liveApplications = liveApplications;
   }
 
 
-  public AccountAnalytics sandboxApplications(Integer sandboxApplications) {
+  public AccountAnalytics sandboxApplications(Long sandboxApplications) {
     
     this.sandboxApplications = sandboxApplications;
     return this;
@@ -167,17 +167,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "2", required = true, value = "Total number of sandbox applications in the account.")
 
-  public Integer getSandboxApplications() {
+  public Long getSandboxApplications() {
     return sandboxApplications;
   }
 
 
-  public void setSandboxApplications(Integer sandboxApplications) {
+  public void setSandboxApplications(Long sandboxApplications) {
     this.sandboxApplications = sandboxApplications;
   }
 
 
-  public AccountAnalytics campaigns(Integer campaigns) {
+  public AccountAnalytics campaigns(Long campaigns) {
     
     this.campaigns = campaigns;
     return this;
@@ -189,17 +189,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "35", required = true, value = "Total number of campaigns in the account.")
 
-  public Integer getCampaigns() {
+  public Long getCampaigns() {
     return campaigns;
   }
 
 
-  public void setCampaigns(Integer campaigns) {
+  public void setCampaigns(Long campaigns) {
     this.campaigns = campaigns;
   }
 
 
-  public AccountAnalytics activeCampaigns(Integer activeCampaigns) {
+  public AccountAnalytics activeCampaigns(Long activeCampaigns) {
     
     this.activeCampaigns = activeCampaigns;
     return this;
@@ -211,17 +211,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "15", required = true, value = "Total number of active campaigns in the account.")
 
-  public Integer getActiveCampaigns() {
+  public Long getActiveCampaigns() {
     return activeCampaigns;
   }
 
 
-  public void setActiveCampaigns(Integer activeCampaigns) {
+  public void setActiveCampaigns(Long activeCampaigns) {
     this.activeCampaigns = activeCampaigns;
   }
 
 
-  public AccountAnalytics liveActiveCampaigns(Integer liveActiveCampaigns) {
+  public AccountAnalytics liveActiveCampaigns(Long liveActiveCampaigns) {
     
     this.liveActiveCampaigns = liveActiveCampaigns;
     return this;
@@ -233,17 +233,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "10", required = true, value = "Total number of active campaigns in live applications in the account.")
 
-  public Integer getLiveActiveCampaigns() {
+  public Long getLiveActiveCampaigns() {
     return liveActiveCampaigns;
   }
 
 
-  public void setLiveActiveCampaigns(Integer liveActiveCampaigns) {
+  public void setLiveActiveCampaigns(Long liveActiveCampaigns) {
     this.liveActiveCampaigns = liveActiveCampaigns;
   }
 
 
-  public AccountAnalytics coupons(Integer coupons) {
+  public AccountAnalytics coupons(Long coupons) {
     
     this.coupons = coupons;
     return this;
@@ -255,17 +255,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "850", required = true, value = "Total number of coupons in the account.")
 
-  public Integer getCoupons() {
+  public Long getCoupons() {
     return coupons;
   }
 
 
-  public void setCoupons(Integer coupons) {
+  public void setCoupons(Long coupons) {
     this.coupons = coupons;
   }
 
 
-  public AccountAnalytics activeCoupons(Integer activeCoupons) {
+  public AccountAnalytics activeCoupons(Long activeCoupons) {
     
     this.activeCoupons = activeCoupons;
     return this;
@@ -277,17 +277,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "650", required = true, value = "Total number of active coupons in the account.")
 
-  public Integer getActiveCoupons() {
+  public Long getActiveCoupons() {
     return activeCoupons;
   }
 
 
-  public void setActiveCoupons(Integer activeCoupons) {
+  public void setActiveCoupons(Long activeCoupons) {
     this.activeCoupons = activeCoupons;
   }
 
 
-  public AccountAnalytics expiredCoupons(Integer expiredCoupons) {
+  public AccountAnalytics expiredCoupons(Long expiredCoupons) {
     
     this.expiredCoupons = expiredCoupons;
     return this;
@@ -299,17 +299,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "200", required = true, value = "Total number of expired coupons in the account.")
 
-  public Integer getExpiredCoupons() {
+  public Long getExpiredCoupons() {
     return expiredCoupons;
   }
 
 
-  public void setExpiredCoupons(Integer expiredCoupons) {
+  public void setExpiredCoupons(Long expiredCoupons) {
     this.expiredCoupons = expiredCoupons;
   }
 
 
-  public AccountAnalytics referralCodes(Integer referralCodes) {
+  public AccountAnalytics referralCodes(Long referralCodes) {
     
     this.referralCodes = referralCodes;
     return this;
@@ -321,17 +321,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "500", required = true, value = "Total number of referral codes in the account.")
 
-  public Integer getReferralCodes() {
+  public Long getReferralCodes() {
     return referralCodes;
   }
 
 
-  public void setReferralCodes(Integer referralCodes) {
+  public void setReferralCodes(Long referralCodes) {
     this.referralCodes = referralCodes;
   }
 
 
-  public AccountAnalytics activeReferralCodes(Integer activeReferralCodes) {
+  public AccountAnalytics activeReferralCodes(Long activeReferralCodes) {
     
     this.activeReferralCodes = activeReferralCodes;
     return this;
@@ -343,17 +343,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "100", required = true, value = "Total number of active referral codes in the account.")
 
-  public Integer getActiveReferralCodes() {
+  public Long getActiveReferralCodes() {
     return activeReferralCodes;
   }
 
 
-  public void setActiveReferralCodes(Integer activeReferralCodes) {
+  public void setActiveReferralCodes(Long activeReferralCodes) {
     this.activeReferralCodes = activeReferralCodes;
   }
 
 
-  public AccountAnalytics expiredReferralCodes(Integer expiredReferralCodes) {
+  public AccountAnalytics expiredReferralCodes(Long expiredReferralCodes) {
     
     this.expiredReferralCodes = expiredReferralCodes;
     return this;
@@ -365,17 +365,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "400", required = true, value = "Total number of expired referral codes in the account.")
 
-  public Integer getExpiredReferralCodes() {
+  public Long getExpiredReferralCodes() {
     return expiredReferralCodes;
   }
 
 
-  public void setExpiredReferralCodes(Integer expiredReferralCodes) {
+  public void setExpiredReferralCodes(Long expiredReferralCodes) {
     this.expiredReferralCodes = expiredReferralCodes;
   }
 
 
-  public AccountAnalytics activeRules(Integer activeRules) {
+  public AccountAnalytics activeRules(Long activeRules) {
     
     this.activeRules = activeRules;
     return this;
@@ -387,17 +387,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "35", required = true, value = "Total number of active rules in the account.")
 
-  public Integer getActiveRules() {
+  public Long getActiveRules() {
     return activeRules;
   }
 
 
-  public void setActiveRules(Integer activeRules) {
+  public void setActiveRules(Long activeRules) {
     this.activeRules = activeRules;
   }
 
 
-  public AccountAnalytics users(Integer users) {
+  public AccountAnalytics users(Long users) {
     
     this.users = users;
     return this;
@@ -409,17 +409,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(required = true, value = "Total number of users in the account.")
 
-  public Integer getUsers() {
+  public Long getUsers() {
     return users;
   }
 
 
-  public void setUsers(Integer users) {
+  public void setUsers(Long users) {
     this.users = users;
   }
 
 
-  public AccountAnalytics roles(Integer roles) {
+  public AccountAnalytics roles(Long roles) {
     
     this.roles = roles;
     return this;
@@ -431,17 +431,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "10", required = true, value = "Total number of roles in the account.")
 
-  public Integer getRoles() {
+  public Long getRoles() {
     return roles;
   }
 
 
-  public void setRoles(Integer roles) {
+  public void setRoles(Long roles) {
     this.roles = roles;
   }
 
 
-  public AccountAnalytics customAttributes(Integer customAttributes) {
+  public AccountAnalytics customAttributes(Long customAttributes) {
     
     this.customAttributes = customAttributes;
     return this;
@@ -453,17 +453,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "18", required = true, value = "Total number of custom attributes in the account.")
 
-  public Integer getCustomAttributes() {
+  public Long getCustomAttributes() {
     return customAttributes;
   }
 
 
-  public void setCustomAttributes(Integer customAttributes) {
+  public void setCustomAttributes(Long customAttributes) {
     this.customAttributes = customAttributes;
   }
 
 
-  public AccountAnalytics webhooks(Integer webhooks) {
+  public AccountAnalytics webhooks(Long webhooks) {
     
     this.webhooks = webhooks;
     return this;
@@ -475,17 +475,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "2", required = true, value = "Total number of webhooks in the account.")
 
-  public Integer getWebhooks() {
+  public Long getWebhooks() {
     return webhooks;
   }
 
 
-  public void setWebhooks(Integer webhooks) {
+  public void setWebhooks(Long webhooks) {
     this.webhooks = webhooks;
   }
 
 
-  public AccountAnalytics loyaltyPrograms(Integer loyaltyPrograms) {
+  public AccountAnalytics loyaltyPrograms(Long loyaltyPrograms) {
     
     this.loyaltyPrograms = loyaltyPrograms;
     return this;
@@ -497,17 +497,17 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "5", required = true, value = "Total number of all loyalty programs in the account.")
 
-  public Integer getLoyaltyPrograms() {
+  public Long getLoyaltyPrograms() {
     return loyaltyPrograms;
   }
 
 
-  public void setLoyaltyPrograms(Integer loyaltyPrograms) {
+  public void setLoyaltyPrograms(Long loyaltyPrograms) {
     this.loyaltyPrograms = loyaltyPrograms;
   }
 
 
-  public AccountAnalytics liveLoyaltyPrograms(Integer liveLoyaltyPrograms) {
+  public AccountAnalytics liveLoyaltyPrograms(Long liveLoyaltyPrograms) {
     
     this.liveLoyaltyPrograms = liveLoyaltyPrograms;
     return this;
@@ -519,12 +519,12 @@ public class AccountAnalytics {
   **/
   @ApiModelProperty(example = "5", required = true, value = "Total number of live loyalty programs in the account.")
 
-  public Integer getLiveLoyaltyPrograms() {
+  public Long getLiveLoyaltyPrograms() {
     return liveLoyaltyPrograms;
   }
 
 
-  public void setLiveLoyaltyPrograms(Integer liveLoyaltyPrograms) {
+  public void setLiveLoyaltyPrograms(Long liveLoyaltyPrograms) {
     this.liveLoyaltyPrograms = liveLoyaltyPrograms;
   }
