+++ src/main/java/one/talon/model/ApplicationEntity.java
@@ -31,10 +31,10 @@ import java.io.IOException;
 public class ApplicationEntity {
   public static final String SERIALIZED_NAME_APPLICATION_ID = "applicationId";
   @SerializedName(SERIALIZED_NAME_APPLICATION_ID)
-  private Integer applicationId;
+  private Long applicationId;
 
 
-  public ApplicationEntity applicationId(Integer applicationId) {
+  public ApplicationEntity applicationId(Long applicationId) {
     
     this.applicationId = applicationId;
     return this;
@@ -46,12 +46,12 @@ public class ApplicationEntity {
   **/
   @ApiModelProperty(example = "322", required = true, value = "The ID of the Application that owns this entity.")
 
-  public Integer getApplicationId() {
+  public Long getApplicationId() {
     return applicationId;
   }
 
 
-  public void setApplicationId(Integer applicationId) {
+  public void setApplicationId(Long applicationId) {
     this.applicationId = applicationId;
   }
